# Cache chrom_info generation when mapping over collections.