@jmchilton 2018-07-20
Failed qunit test unrelated to this PR - I think https://github.com/galaxyproject/galaxy/pull/6520 should fix it.